refactor(persona): move VrmSelect component to shared
