--- mods/persona/management/src/hooks/useClickOutside.ts
@@ -1,1 +1,0 @@
-export { useClickOutside } from "@persona/shared/hooks/useClickOutside";
