--- mods/persona/management/src/hooks/useVrmAssets.ts
@@ -1,1 +1,0 @@
-export { useVrmAssets, type UseVrmAssetsReturn } from "@persona/shared/hooks/useVrmAssets";
